pyproject.toml
[tool.poetry]
name = "onward-pc-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
nextcord = "^2.0.0-alpha.6"
python-dotenv = "^0.19.2"
Flask = "^2.0.2"
"statcord.py-beta" = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
